Add extend and blocks feature, the only missed on tpl class.
It's really simple to use, you can use default.php template for the base template, and example for the home template you need to add only the content block, here the examples:

*default.php*

    <h1>Default template</h1>
    
    <?php Tpl::block('body_content') ?>

*home.php*

    <?php Tpl::extend_start('default') ?>
    
        <?php Tpl::block_start('body_content') ?>
            <h1><?php echo $page->title()->html() ?></h1>
        <?php Tpl::block_end('body_content') ?>
    
    <?php Tpl::extend_end('default') ?>

In this case the block "body_content" replaced from *home.php* template from *default.php* template.

Anyone have more fine idea for this? I hate all the times replace the default.php template with another one and porting some code from default template. It's true we can use the snippet, but for me it's really useless... best way is the block, i take this idea from twig (the best template engine i tried in this years), the extended it's really nice, we can use only one call but we need to add to the end of template, so with start/end it's more clear.