Store safe name in variable
The `f::safeName()` method needs a lot of CPU time, so it should not be executed more times than necessary. On a page with dozens of thumbnails, a PHP profiler told me, that it was responsible for 30-50% of the total page generation time.

If the thumb class is also intended for generating `<picture>` tags with multiple sources of the same image, you should also consider to store every result of the function in a static variable to avoid double-execution of the function.