Accept PHP's syntax for multiple uploaded files
See http://php.net/manual/en/features.file-upload.multiple.php.

To use this PHP feature with this class, the user can create multiple instances of `Upload` in a loop, each with an incremented `index` value.

Also see https://forum.getkirby.com/t/allow-guests-to-create-pages-from-panel-or-frontend/2854/8?u=lukasbestle for the background of this feature.